add error checking

+++ src/main/java/com/nvidia/spark/rapids/jni/DeviceAttr.java
@@ -7,9 +7,9 @@ public class DeviceAttr {
     NativeDepsLoader.loadNativeDeps();
   }
 
-  public static long isIntegratedGPU() {
+  public static int isIntegratedGPU() {
     return isDeviceIntegrated();
   }
 
-  private static native long isDeviceIntegrated();
+  private static native int isDeviceIntegrated();
 }
